Make clippy happy

+++ fazic_lib/src/screen/mod.rs
@@ -53,11 +53,11 @@ pub fn draw_text_buffer(fazic: &mut ::Fazic) {
 }
 
 pub fn line(fazic: &mut ::Fazic, x0: i32, y0: i32, x1: i32, y1: i32, color: u8) {
-    let mut x = x0 as i32;
-    let mut y = y0 as i32;
+    let mut x = x0;
+    let mut y = y0;
     let mut y_longer = false;
-    let mut short_len = y1 as i32 - y;
-    let mut long_len = x1 as i32 - x;
+    let mut short_len = y1 - y;
+    let mut long_len = x1 - x;
 
     if short_len.abs() > long_len.abs() {
         mem::swap(&mut short_len, &mut long_len);
